UE5.4へ移行，バージョン表記をv1.6.19 に更新

+++ Ss6PlayerExamples/Plugins/SpriteStudio6/Source/SpriteStudio6/Private/Render/SsRenderOffScreen.cpp
@@ -208,7 +208,7 @@ namespace
 			case SsBlendType::Add:
 			case SsBlendType::Sub:
 				{
-					OutColorBlend.X = (float)(ColorBlendType + 0.01f);
+					OutColorBlend.X = (float)((int32)ColorBlendType + 0.01f);
 				} break;
 			case SsBlendType::MulAlpha:
 			case SsBlendType::Screen:
